chore: Refactor CLI code and implement directory cleaning prompt and dry run option
